// frontend/src/api/models/registerRequestRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Gestionale Negozio API
 * API del gestionale negozio
 * OpenAPI spec version: 1.0.0
 */

export type RegisterRequestRole = typeof RegisterRequestRole[keyof typeof RegisterRequestRole];


export const RegisterRequestRole = {
  ADMIN: 'ADMIN',
  COMMESSO: 'COMMESSO',
} as const;